Refactor(API): 변경 요구사항에 맞추어 API 입력값 수정
변경된 요구사항에 맞추어 API 입력값 수정

+++ src/main/java/com/project/shoppingmall/controller/chat_retrieve/ChatRetrieveController.java
@@ -67,7 +67,7 @@ public class ChatRetrieveController {
   @PreAuthorize("hasRole('ROLE_MEMBER')")
   public OutputRetrieveChatMessage retrieveChatMessages(
       @Positive @RequestParam("sliceSize") Integer sliceSize,
-      @RequestParam("chatRoomId") Integer chatRoomId,
+      @RequestParam("chatRoomId") Long chatRoomId,
       @Length(min = 1, max = 50) @RequestParam("startChatMessageId") String startChatMessageId) {
     AuthMemberDetail userDetail =
         (AuthMemberDetail) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
